Last brashes

+++ src/main/scala/com/myapp/domain/Protocol.scala
@@ -6,11 +6,11 @@ object Protocol {
 
   type Chunk = mutable.Queue[UserData]
 
-  case class UserData(userId: String, data: String, weight: Float, int1: Int, int2: Long)
+  case class UserData(userId: String, data: String, weight: Double, int1: Int, int2: BigInt)
 
-  case class AggregatedUserData(sumInt2: Long, usersCount: Int, usersTotals: List[UserTotals])
+  case class AggregatedUserData(sumInt2: BigInt, usersCount: Int, usersTotals: List[UserTotals])
 
-  case class UserTotals(userId: String, avgWeight: Float, latestInt1: Int) {
+  case class UserTotals(userId: String, avgWeight: Double, latestInt1: Int) {
     override def toString: String =
       userId + "," + avgWeight + "," + latestInt1
   }
